+++ .gitignore
@@ -10,3 +10,4 @@ receipt.txt
 .idea/shelf/Default_Changelist.xml
 /config.json
 /roles.json.bak
+/stonks.json.bak
