+++ cogs/commands/debug.py
@@ -1,0 +1,43 @@
+import discord
+from discord.ext import commands
+
+from eggbot import host_check
+
+
+class Debug(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
+
+    @commands.command()
+    async def test_args(self, ctx, *args):
+        arghs = args
+        argsleft = len(arghs)
+        emb = discord.Embed(title="Arguments", description="Arguments", color=0x0f88f0)
+        if argsleft == 0:
+            emb.add_field(name="Error", value="No arguments detected", inline=False)
+        else:
+            argNumber = 0
+            while 0 < argsleft:
+                argnotext = str(argNumber + 1)
+                emb.add_field(name="Argument " + argnotext, value=arghs[argNumber], inline=False)
+                argsleft = argsleft - 1
+                argNumber = 1 + argNumber
+            argnotext = str(len(arghs))
+            emb.add_field(name="Total Arguments", value=argnotext, inline=False)
+        await ctx.send(embed=emb)
+
+    @commands.command()
+    async def print_emoji(self, ctx, *args):
+        if host_check(ctx):
+            print(args[0])
+            await ctx.send('Check the console!')
+
+    @commands.command()
+    async def embedTest(self, ctx):
+        embed = discord.Embed(title='Go to YouTube', url='https://www.youtube.com/',
+                              description='New video guys click on the title or click [here](https://www.youtube.com/)')
+        await ctx.send(embed=embed)
+
+
+def setup(bot):
+    bot.add_cog(Debug(bot))
