+++ cogs/commands/settings.py
@@ -1,7 +1,10 @@
 import discord
 from discord.ext import commands
 
-from eggbot import dmLog, logging, audit, deleteLog
+safeguard = True
+botSafeguard = True
+status = "e!help"
+from eggbot import dmLog, logging, audit, deleteLog, host_check, joinArgs, reverseBool
 
 
 def settingCheck(setting):
@@ -15,28 +18,25 @@ class Settings(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    # @commands.command()
-    # async def spam(self, ctx):
-    #     if host_check(ctx):
-    #         global safeguard
-    #         safeguard, state = reverseBool(safeguard)
-    #         await ctx.send("Set spam mode to {}.".format(state.upper()))
-# TODO fix these
-    # @commands.command()
-    # async def botSpam(self, ctx):
-    #     if host_check(ctx):
-    #         global botSafeguard
-    #         botSafeguard, state = reverseBool(botSafeguard)
-    #         await ctx.send("Set bot message processing to {}.".format(state.upper()))
+    @commands.command()
+    async def spam(self, ctx):
+        if host_check(ctx):
+            self.bot.safeguard, state = reverseBool(self.bot.safeguard)
+            await ctx.send("Set spam mode to {}.".format(state.upper()))
 
-    # @bot.command()
-    # async def setStatus(ctx, *args):
-    #     if host_check(ctx):
-    #         global status
-    #         args = list(args)
-    #         status = joinArgs(args)
-    #         await bot.change_presence(activity=discord.Game(name=status))
-    #         await ctx.send('Status set to "{s}".'.format(s=status))
+    @commands.command()
+    async def botSpam(self, ctx):
+        if host_check(ctx):
+            self.bot.botSafeguard, state = reverseBool(self.bot.botSafeguard)
+            await ctx.send("Set bot message processing to {}.".format(state.upper()))
+
+    @commands.command()
+    async def setStatus(self, ctx, *args):
+        if host_check(ctx):
+            args = list(args)
+            self.bot.status = joinArgs(args)
+            await self.bot.change_presence(activity=discord.Game(name=self.bot.status))
+            await ctx.send('Status set to "{s}".'.format(s=self.bot.status))
 
     @commands.command()
     async def settings(self, ctx):
