+++ roles.json
@@ -1,1 +1,1 @@
-{"reactions": {"708056795523973292": {"\ud83d\udcb0": {"role": 706296742114754670, "addMessage": "You will now be pinged when a major announcement appears in <#705235263428886560>", "removeMessage": "You will no longer be pinged when a major announcement appears in <#705235263428886560>"}, "<:earth:708046023750320179>": {"role": 708042351037120512, "addMessage": "You will now be pinged when there is an announcement in Earth 2.0.", "removeMessage": "You will no longer be pinged when there is an announcement in Earth 2.0."}}, "708435235569664078": {"<:ooo:704401624289771611>": {"role": 708421333888794644, "addMessage": "You will now be pinged when a major announcement appears in <#706645368678645831>", "removeMessage": "You will no longer be pinged when a major announcement appears in <#706645368678645831>"}}, "722963178081746964": {"<:egghead:698281894441123990>": {"role": 722962992215359538}}, "723218448502489258": {"<a:Duck:718494534111068183>": {"role": 717393407663669330}}, "724616552317452368": {"\ud83d\udcfa": {"role": 721479066716209224}, "\ud83c\udd93": {"role": 718564157242343485}}, "725431730303008859": {"<:egghead:698281894441123990>": {"role": 725428970694115391}}, "729077156298096743": {"<:helloThere:719939637883306044>": {"role": 729075315569197138}}, "731208664777752596": {"\ud83d\udcd6": {"role": 731208430349713408}}, "731209276122726430": {"\ud83d\udd14": {"role": 731208902665961473}}, "731247163291467889": {"\ud83d\udcd6": {"role": 731208430349713408}}, "674430988876775426": 709561341236740198, "702578905118146641": 731337129178693633, "737361020170928189": {"\ud83d\udcdd": {"role": 731208430349713408}}}, "join": {"713488900890296431": 721450592559693906}}
+{"reactions": {"708056795523973292": {"\ud83d\udcb0": {"role": 706296742114754670, "addMessage": "You will now be pinged when a major announcement appears in <#705235263428886560>", "removeMessage": "You will no longer be pinged when a major announcement appears in <#705235263428886560>"}, "<:earth:708046023750320179>": {"role": 708042351037120512, "addMessage": "You will now be pinged when there is an announcement in Earth 2.0.", "removeMessage": "You will no longer be pinged when there is an announcement in Earth 2.0."}}, "708435235569664078": {"<:ooo:704401624289771611>": {"role": 708421333888794644, "addMessage": "You will now be pinged when a major announcement appears in <#706645368678645831>", "removeMessage": "You will no longer be pinged when a major announcement appears in <#706645368678645831>"}}, "722963178081746964": {"<:egghead:698281894441123990>": {"role": 722962992215359538}}, "723218448502489258": {"<a:Duck:718494534111068183>": {"role": 717393407663669330}}, "724616552317452368": {"\ud83d\udcfa": {"role": 721479066716209224}, "\ud83c\udd93": {"role": 718564157242343485}}, "725431730303008859": {"<:egghead:698281894441123990>": {"role": 725428970694115391}}, "729077156298096743": {"<:helloThere:719939637883306044>": {"role": 729075315569197138}}, "731208664777752596": {"\ud83d\udcd6": {"role": 731208430349713408}}, "731209276122726430": {"\ud83d\udd14": {"role": 731208902665961473}}, "731247163291467889": {"\ud83d\udcd6": {"role": 731208430349713408}}, "674430988876775426": 709561341236740198, "702578905118146641": 731337129178693633, "737361020170928189": {"\ud83d\udcdd": {"role": 731208430349713408}}}, "join": {"713488900890296431": 721450592559693906, "726523476306952293": 726524871428800575}}
